testing/web-platform/mozilla/tests/focus/Range_setEnd.html

<!doctype html>
<meta charset=utf-8>
<title>focus move tests caused by a call of Range.setEnd(), Range.setEndAfter() and Range.setEndBefore()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<p id="staticBefore">static text</p>
<div id="editor" contenteditable><p>content of editor</p></div>
<div id="outerEditor" contenteditable
><p>content of outer editor</p><div id="staticInEditor" contenteditable="false"
><p>static content of outer editor</p><div id="innerEditor" contenteditable
><p>content of inner editor</p></div></div></div>
<p id="staticAfter">static text</p>
<p><a id="anchor" href="about:blank">anchor</a></p>
<script>
"use strict";

var staticBefore = {
    element: document.getElementById("staticBefore"),
    textNode: document.getElementById("staticBefore").firstChild,
    textLength: document.getElementById("staticBefore").firstChild.length
};
var editor = {
    element: document.getElementById("editor"),
    textNode: document.getElementById("editor").firstChild.firstChild,
    textLength: document.getElementById("editor").firstChild.firstChild.length
};
var outerEditor = {
    element: document.getElementById("outerEditor"),
    textNode: document.getElementById("outerEditor").firstChild.firstChild,
    textLength: document.getElementById("outerEditor").firstChild.firstChild.length
};
var staticInEditor = {
    element: document.getElementById("staticInEditor"),
    textNode: document.getElementById("staticInEditor").firstChild,
    textLength: document.getElementById("staticInEditor").firstChild.length
};
var innerEditor = {
    element: document.getElementById("innerEditor"),
    textNode: document.getElementById("innerEditor").firstChild.firstChild,
    textLength: document.getElementById("innerEditor").firstChild.firstChild.length
};
var staticAfter = {
    element: document.getElementById("staticAfter"),
    textNode: document.getElementById("staticAfter").firstChild,
    textLength: document.getElementById("staticAfter").firstChild.length
};
var anchor = {
    element: document.getElementById("anchor"),
    textNode: document.getElementById("anchor").firstChild,
    textLength: document.getElementById("anchor").firstChild.length
};

function resetFocusAndSelectionRange(aFocus)
{
    document.getSelection().removeAllRanges();
    if (document.activeElement) {
        document.activeElement.blur();
    }
    if (aFocus) {
        aFocus.element.focus();
        document.getSelection().collapse(aFocus.textNode, 0);
    } else {
        document.getSelection().collapse(staticBefore.textNode, 0);
    }
}

function setEnd(aNode, aOffset)
{
    document.getSelection().getRangeAt(0).setEnd(aNode, aOffset);
}

function setEndBefore(aNode, aOffset)
{
    document.getSelection().getRangeAt(0).setEndBefore(aNode);
}

function setEndAfter(aNode, aOffset)
{
    document.getSelection().getRangeAt(0).setEndAfter(aNode);
}

// Range.setEnd*() should work same as collapse if specified end position is before its start.
[{ func: setEnd, doingDescription: "Range.setEnd()" },
 { func: setEndBefore, doingDescription: "Range.setEndBefore()" }].forEach((aTest, aIndex, aArray)=>{
    test(function() {
        resetFocusAndSelectionRange(editor);
        document.getSelection().selectAllChildren(editor.textNode);
        aTest.func(staticBefore.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'staticBefore' (before the selection) when active element is 'editor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().selectAllChildren(outerEditor.textNode);
        aTest.func(editor.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'editor' (before the selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().selectAllChildren(staticInEditor.textNode);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (before the selection) when active element is the <body>");
    test(function() {
        resetFocusAndSelectionRange(innerEditor);
        document.getSelection().selectAllChildren(innerEditor.textNode);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (before the selection) when active element is 'innerEditor'");
    test(function() {
        resetFocusAndSelectionRange(anchor);
        document.getSelection().selectAllChildren(anchor.textNode);
        aTest.func(staticAfter.textNode, 0);
        assert_equals(document.activeElement, anchor.element);
    }, "Active element should be 'anchor' after " + aTest.doingDescription + " with start of the first text node of 'staticAfter' (before the selection) when active element is 'anchor'");

    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().collapse(staticBefore.textNode, staticBefore.textLength);
        aTest.func(staticBefore.textNode, 0);
        assert_equals(document.activeElement, document.body);
    }, "Active element should be the <body> after " + aTest.doingDescription + " with start of the first text node of 'staticBefore' (before the collapsed selection) when active element is the <body>");
    test(function() {
        resetFocusAndSelectionRange(editor);
        document.getSelection().collapse(editor.textNode, editor.textLength);
        aTest.func(staticBefore.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'staticBefore' (before the collapsed selection) when active element is 'editor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().collapse(outerEditor.textNode, outerEditor.textLength);
        aTest.func(editor.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'editor' (before the collapsed selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().collapse(staticInEditor.textNode, staticInEditor.textLength);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (before the collapsed selection) when active element is the <body> and selection is in 'staticInEditor'");
    test(function() {
        resetFocusAndSelectionRange(innerEditor);
        document.getSelection().collapse(innerEditor.textNode, innerEditor.textLength);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (before the collapsed selection) when active element is 'innerEditor'");
    test(function() {
        resetFocusAndSelectionRange(anchor);
        document.getSelection().collapse(anchor.textNode, anchor.textLength);
        aTest.func(staticAfter.textNode, 0);
        assert_equals(document.activeElement, anchor.element);
    }, "Active element should be 'anchor' after " + aTest.doingDescription + " with start of the first text node of 'staticAfter' (before the collapsed selection) when active element is 'anchor'");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().collapse(anchor.textNode, anchor.textLength);
        aTest.func(anchor.textNode, 0);
        assert_equals(document.activeElement, document.body);
    }, "Active element should be the <body> after " + aTest.doingDescription + " with start of the first text node of 'anchor' (before the collapsed selection) when active element is the <body>");

    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().setBaseAndExtent(staticBefore.textNode, staticBefore.textLength,
                                                 editor.textNode, editor.textLength);
        aTest.func(staticBefore.textNode, 0);
        assert_equals(document.activeElement, document.body);
    }, "Active element should be the <body> after " + aTest.doingDescription + " with start of the first text node of 'staticBefore' (before the selection, between end of the first text node of 'staticBefore' and end of the first text node of 'editor') when active element is the <body>");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().setBaseAndExtent(editor.textNode, editor.textLength,
                                                 outerEditor.textNode, outerEditor.textLength);
        aTest.func(editor.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'editor' (before the selection, between end of the first text node of 'editor' and end of the first text node of 'outerEditor') when active element is the <body>");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().setBaseAndExtent(outerEditor.textNode, outerEditor.textLength,
                                                 innerEditor.textNode, innerEditor.textLength);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (before the selection, between end of the first text node of 'outerEditor' and end of the first text node of 'innerEditor') when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().setBaseAndExtent(staticInEditor.textNode, staticInEditor.textLength,
                                                 innerEditor.textNode, innerEditor.textLength);
        aTest.func(staticInEditor.textNode, 0);
        assert_equals(document.activeElement, document.body);
    }, "Active element should be the <body> after " + aTest.doingDescription + " with start of the first text node of 'staticInEditor' (before the selection, between end of the first text node of 'staticInEditor' and end of the first text node of 'innerEditor') when active element is the <body>");
});

test(function() {
    resetFocusAndSelectionRange();
    document.getSelection().selectAllChildren(staticAfter.textNode);
    setEndAfter(innerEditor.textNode);
    assert_equals(document.activeElement, innerEditor.element);
}, "Active element should be 'innerEditor' after Selection.setEndAfter() with the first text node of 'innerEditor' (before the selection) when active element is the <body> and selection is in 'staticAfter'");
test(function() {
    resetFocusAndSelectionRange(innerEditor);
    document.getSelection().selectAllChildren(innerEditor.textNode);
    setEndAfter(staticInEditor.textNode);
    assert_equals(document.activeElement, innerEditor.element);
}, "Active element should be 'innerEditor' after Selection.setEndAfter() with the first text node of 'staticInEditor' (before the selection) when active element is 'innerEditor'");
test(function() {
    resetFocusAndSelectionRange();
    document.getSelection().selectAllChildren(staticInEditor.textNode);
    setEndAfter(outerEditor.textNode);
    assert_equals(document.activeElement, outerEditor.element);
}, "Active element should be 'outerEditor' after Selection.setEndAfter() with the first text node of 'outerEditor' (before the selection) when active element is the <body> and selection is in 'staticInEditor'");
test(function() {
    resetFocusAndSelectionRange(innerEditor);
    document.getSelection().selectAllChildren(innerEditor.textNode);
    setEndAfter(outerEditor.textNode);
    assert_equals(document.activeElement, outerEditor.element);
}, "Active element should be 'outerEditor' after Selection.setEndAfter() with the first text node of 'outerEditor' (before the selection) when active element is 'innerEditor'");
test(function() {
    resetFocusAndSelectionRange(outerEditor);
    document.getSelection().selectAllChildren(outerEditor.textNode);
    setEndAfter(editor.textNode);
    assert_equals(document.activeElement, editor.element);
}, "Active element should be 'editor' after Selection.setEndAfter() with the first text node of 'editor' (before the selection) when active element is 'outerEditor'");
test(function() {
    resetFocusAndSelectionRange(editor);
    document.getSelection().selectAllChildren(editor.textNode);
    setEndAfter(staticBefore.textNode);
    assert_equals(document.activeElement, editor.element);
}, "Active element should be 'editor' after Selection.setEndAfter() with the first text node of 'staticBefore' (before the selection) when active element is 'editor'");
test(function() {
    resetFocusAndSelectionRange(anchor);
    document.getSelection().selectAllChildren(anchor.textNode);
    setEndAfter(staticBefore.textNode);
    assert_equals(document.activeElement, anchor.element);
}, "Active element should be 'anchor' after Selection.setEndAfter() with the first text node of 'staticBefore' (before the selection) when active element is 'anchor'");

// Range.setEnd*() should blur focused editing host when it expands selection to outside of it.
[{ func: setEnd, doingDescription: "Range.setEnd()" },
 { func: setEndAfter, doingDescription: "Range.setEndAfter()" },
 { func: setEndBefore, doingDescription: "Range.setEndBefore()" }].forEach((aTest, aIndex, aArray)=>{
    test(function() {
        resetFocusAndSelectionRange(editor);
        document.getSelection().collapse(editor.textNode, 0);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (after end of the collapsed selection) when active element is 'editor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().collapse(outerEditor.textNode, 0);
        aTest.func(staticInEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (after end of the collapsed selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().collapse(outerEditor.textNode, 0);
        aTest.func(innerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'innerEditor' (after end of the collapsed selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().collapse(outerEditor.textNode, 0);
        aTest.func(staticAfter.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'staticAfter' (after end of the collapsed selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange(editor);
        document.getSelection().selectAllChildren(editor.textNode);
        aTest.func(outerEditor.textNode, 0);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (after end of the selection) when active element is 'editor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().selectAllChildren(outerEditor.textNode);
        aTest.func(staticInEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'outerEditor' (after end of the selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().selectAllChildren(outerEditor.textNode);
        aTest.func(innerEditor.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'innerEditor' (after end of the selection) when active element is 'outerEditor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().selectAllChildren(outerEditor.textNode);
        aTest.func(staticAfter.textNode, 0);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with start of the first text node of 'staticAfter' (after end of the selection) when active element is 'outerEditor'");
});

// Range.setEnd*() should move focus to an editing host when the range is shrunken into it.
[{ func: setEnd, doingDescription: "Range.setEnd()" },
 { func: setEndAfter, doingDescription: "Range.setEndAfter()" }].forEach((aTest, aIndex, aArray)=>{
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().setBaseAndExtent(editor.textNode, 0,
                                                 outerEditor.textNode, outerEditor.textLength);
        aTest.func(editor.textNode, editor.textLength);
        assert_equals(document.activeElement, editor.element);
    }, "Active element should be 'editor' after " + aTest.doingDescription + " with end of the first text node of 'editor' (shrunken into 'editor') when active element is the <body>");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().setBaseAndExtent(outerEditor.textNode, 0,
                                                 staticInEditor.textNode, staticInEditor.textLength);
        aTest.func(outerEditor.textNode, outerEditor.textLength);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with end of the first text node of 'outerEditor' (shrunken into 'outerEditor') when active element is 'outerEditor' and selection end is in 'staticInEditor'");
    test(function() {
        resetFocusAndSelectionRange(outerEditor);
        document.getSelection().setBaseAndExtent(outerEditor.textNode, 0,
                                                 innerEditor.textNode, innerEditor.textLength);
        aTest.func(outerEditor.textNode, outerEditor.textLength);
        assert_equals(document.activeElement, outerEditor.element);
    }, "Active element should be 'outerEditor' after " + aTest.doingDescription + " with end of the first text node of 'outerEditor' (shrunken into 'outerEditor') when active element is 'outerEditor' and selection end is in 'innerEditor'");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().setBaseAndExtent(innerEditor.textNode, 0,
                                                 staticAfter.textNode, staticAfter.textLength);
        aTest.func(innerEditor.textNode, innerEditor.textLength);
        assert_equals(document.activeElement, innerEditor.element);
    }, "Active element should be 'innerEditor' after " + aTest.doingDescription + " with end of the first text node of 'innerEditor' (shrunken into 'innerEditor') when active element is 'innerEditor' and selection end is in 'staticAfter'");
    test(function() {
        resetFocusAndSelectionRange();
        document.getSelection().setBaseAndExtent(staticInEditor.textNode, 0,
                                                 innerEditor.textNode, innerEditor.textLength);
        aTest.func(staticInEditor.textNode, staticInEditor.textLength);
        assert_equals(document.activeElement, document.body);
    }, "Active element should be the <body> after " + aTest.doingDescription + " with end of the first text node of 'staticInEditor' (shrunken into 'staticInEditor') when active element is the <body>");
});
</script>
